+++ app/utils/auth.ts
